move helpers to techdocs-react
