+++ plugins/techdocs/src/index.ts
@@ -46,7 +46,6 @@ export {
   LegacyEmbeddedDocsRouter as EmbeddedDocsRouter,
   Router,
 } from './Router';
-export { buildTechDocsURL, getEntityRootTechDocsPath } from './helpers';
 
 export type { TechDocsSearchResultListItemProps } from './search/components/TechDocsSearchResultListItem';
 
@@ -70,5 +69,3 @@ export type {
 };
 
 export * from './overridableComponents';
-
-export type { TechDocsRouteFunc } from './helpers';
